psp-8784 do not display cancel modal unless changes have been made to lease screen.
