+++ source/frontend/src/features/leases/add/AddLeaseForm.tsx
@@ -2,12 +2,14 @@ import { Formik, FormikHelpers, FormikProps } from 'formik';
 import styled from 'styled-components';
 
 import { IMapProperty } from '@/components/propertySelector/models';
+import * as API from '@/constants/API';
+import useLookupCodeHelpers from '@/hooks/useLookupCodeHelpers';
 
 import { FormLeaseProperty, getDefaultFormLease, LeaseFormModel } from '../models';
 import LeasePropertySelector from '../shared/propertyPicker/LeasePropertySelector';
 import { AddLeaseYupSchema } from './AddLeaseYupSchema';
 import AdministrationSubForm from './AdministrationSubForm';
-import ConsultationSubForm from './ConsultationSubForm';
+import ConsultationSubForm, { getConsultations } from './ConsultationSubForm';
 import LeaseDetailSubForm from './LeaseDetailSubForm';
 import DocumentationSubForm from './ReferenceSubForm';
 
@@ -26,6 +28,10 @@ const AddLeaseForm: React.FunctionComponent<React.PropsWithChildren<IAddLeaseFor
   propertyInfo,
 }) => {
   const defaultFormLease = getDefaultFormLease();
+  const apiFormLease = LeaseFormModel.toApi(defaultFormLease);
+  const { getByType } = useLookupCodeHelpers();
+  const consultationTypes = getByType(API.CONSULTATION_TYPES);
+  apiFormLease.consultations = getConsultations(apiFormLease, consultationTypes);
   if (propertyInfo) {
     defaultFormLease.properties = [];
     defaultFormLease.properties.push(FormLeaseProperty.fromMapProperty(propertyInfo));
@@ -44,7 +50,7 @@ const AddLeaseForm: React.FunctionComponent<React.PropsWithChildren<IAddLeaseFor
       <Formik<LeaseFormModel>
         enableReinitialize
         innerRef={formikRef}
-        initialValues={defaultFormLease}
+        initialValues={LeaseFormModel.fromApi(apiFormLease)}
         validationSchema={AddLeaseYupSchema}
         onSubmit={handleSubmit}
       >
