+++ source/frontend/src/features/leases/add/ConsultationSubForm.tsx
@@ -6,9 +6,11 @@ import { Section } from '@/components/common/Section/Section';
 import { SectionField } from '@/components/common/Section/SectionField';
 import * as API from '@/constants/API';
 import useLookupCodeHelpers from '@/hooks/useLookupCodeHelpers';
-import useDeepCompareEffect from '@/hooks/util/useDeepCompareEffect';
+import { ApiGen_Concepts_ConsultationLease } from '@/models/api/generated/ApiGen_Concepts_ConsultationLease';
+import { ApiGen_Concepts_Lease } from '@/models/api/generated/ApiGen_Concepts_Lease';
+import { getEmptyBaseAudit } from '@/models/defaultInitializers';
 
-import { FormLeaseConsultation, LeaseFormModel } from '../models';
+import { LeaseFormModel } from '../models';
 
 export interface IConsultationSubFormProps {
   formikProps: FormikProps<LeaseFormModel>;
@@ -17,39 +19,9 @@ export interface IConsultationSubFormProps {
 const ConsultationSubForm: React.FunctionComponent<
   React.PropsWithChildren<IConsultationSubFormProps>
 > = ({ formikProps }) => {
-  const { values, setFieldValue } = formikProps;
-  const { consultations } = values;
-  const { getOptionsByType, getByType } = useLookupCodeHelpers();
-  const consultationTypes = getByType(API.CONSULTATION_TYPES);
+  const { values } = formikProps;
+  const { getOptionsByType } = useLookupCodeHelpers();
   const consultationStatusTypes = getOptionsByType(API.CONSULTATION_STATUS_TYPES);
-
-  useDeepCompareEffect(() => {
-    // Not all consultations might be coming from the backend. Add the ones missing.
-    if (consultations.length !== consultationTypes.length) {
-      const newConsultations: FormLeaseConsultation[] = [];
-
-      consultationTypes.forEach(consultationType => {
-        const newConsultation = FormLeaseConsultation.fromApiLookup(
-          values.id || 0,
-          consultationType,
-        );
-
-        // If there is a consultation with the type, set the status to the existing one
-        const existingConsultation = consultations.find(
-          consultation => consultation.consultationType === consultationType.id,
-        );
-        if (existingConsultation !== undefined) {
-          newConsultation.id = existingConsultation.id;
-          newConsultation.consultationStatusType = existingConsultation.consultationStatusType;
-          newConsultation.consultationStatusTypeDescription =
-            existingConsultation.consultationStatusTypeDescription;
-          newConsultation.rowVersion = existingConsultation.rowVersion;
-        }
-        newConsultations.push(newConsultation);
-      });
-      setFieldValue('consultations', newConsultations);
-    }
-  }, [consultationTypes, consultations, setFieldValue, values.id]);
 
   return (
     <Section header="Consultation" isCollapsable initiallyExpanded>
@@ -82,6 +54,46 @@ const ConsultationSubForm: React.FunctionComponent<
       ))}
     </Section>
   );
+};
+
+export const getConsultations = (lease: ApiGen_Concepts_Lease, consultationTypes) => {
+  if (lease.consultations?.length !== consultationTypes.length) {
+    const newConsultations: ApiGen_Concepts_ConsultationLease[] = [];
+
+    consultationTypes.forEach(consultationType => {
+      const newConsultation: ApiGen_Concepts_ConsultationLease = {
+        id: 0,
+        parentLeaseId: lease.id || 0,
+        consultationType: {
+          id: consultationType.id.toString(),
+          description: consultationType.name,
+          displayOrder: null,
+          isDisabled: false,
+        },
+        consultationStatusType: {
+          id: 'UNKNOWN',
+          description: 'Unknown',
+          displayOrder: null,
+          isDisabled: false,
+        },
+        otherDescription: null,
+        ...getEmptyBaseAudit(0),
+      };
+
+      // If there is a consultation with the type, set the status to the existing one
+      const existingConsultation = lease.consultations?.find(
+        consultation => consultation.consultationType?.id === consultationType.id,
+      );
+      if (existingConsultation !== undefined) {
+        newConsultation.id = existingConsultation.id;
+        newConsultation.consultationStatusType = existingConsultation.consultationStatusType;
+        newConsultation.rowVersion = existingConsultation.rowVersion;
+      }
+      newConsultations.push(newConsultation);
+    });
+    return newConsultations;
+  }
+  return [];
 };
 
 export default ConsultationSubForm;
