+++ source/frontend/src/features/leases/detail/LeasePages/details/DetailConsultation.tsx
@@ -2,11 +2,8 @@ import { useFormikContext } from 'formik';
 
 import { Section } from '@/components/common/Section/Section';
 import { SectionField } from '@/components/common/Section/SectionField';
-import * as API from '@/constants/API';
-import useLookupCodeHelpers from '@/hooks/useLookupCodeHelpers';
 import { ApiGen_Concepts_ConsultationLease } from '@/models/api/generated/ApiGen_Concepts_ConsultationLease';
 import { ApiGen_Concepts_Lease } from '@/models/api/generated/ApiGen_Concepts_Lease';
-import { getEmptyBaseAudit } from '@/models/defaultInitializers';
 import { exists } from '@/utils/utils';
 
 export interface IDetailConsultationProps {
@@ -20,48 +17,7 @@ export interface IDetailConsultationProps {
 export const DetailConsultation: React.FunctionComponent<
   React.PropsWithChildren<IDetailConsultationProps>
 > = () => {
-  const { values, setFieldValue } = useFormikContext<ApiGen_Concepts_Lease>();
-
-  const { getByType } = useLookupCodeHelpers();
-  const consultationTypes = getByType(API.CONSULTATION_TYPES);
-
-  // Not all consultations might be coming from the backend. Add the ones missing.
-  if (values.consultations?.length !== consultationTypes.length) {
-    const newConsultations: ApiGen_Concepts_ConsultationLease[] = [];
-
-    consultationTypes.forEach(consultationType => {
-      const newConsultation: ApiGen_Concepts_ConsultationLease = {
-        id: 0,
-        parentLeaseId: values.id || 0,
-        consultationType: {
-          id: consultationType.id.toString(),
-          description: consultationType.name,
-          displayOrder: null,
-          isDisabled: false,
-        },
-        consultationStatusType: {
-          id: 'UNKNOWN',
-          description: 'Unknown',
-          displayOrder: null,
-          isDisabled: false,
-        },
-        otherDescription: null,
-        ...getEmptyBaseAudit(0),
-      };
-
-      // If there is a consultation with the type, set the status to the existing one
-      const existingConsultation = values.consultations?.find(
-        consultation => consultation.consultationType?.id === consultationType.id,
-      );
-      if (existingConsultation !== undefined) {
-        newConsultation.id = existingConsultation.id;
-        newConsultation.consultationStatusType = existingConsultation.consultationStatusType;
-        newConsultation.rowVersion = existingConsultation.rowVersion;
-      }
-      newConsultations.push(newConsultation);
-    });
-    setFieldValue('consultations', newConsultations);
-  }
+  const { values } = useFormikContext<ApiGen_Concepts_Lease>();
 
   const generateLabel = (consultation: ApiGen_Concepts_ConsultationLease): string => {
     let label = consultation.consultationType?.description || '';
